src/pyproject.toml
# SPDX-FileCopyrightText: Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
# SPDX-License-Identifier: LicenseRef-NvidiaProprietary
#
# NVIDIA CORPORATION, its affiliates and licensors retain all intellectual
# property and proprietary rights in and to this material, related
# documentation and any modifications thereto. Any use, reproduction,
# disclosure or distribution of this material and related documentation
# without an express license agreement from NVIDIA CORPORATION or
# its affiliates is strictly prohibited.

[tool.poetry]
name = "dfm-client"
version = "1.5.0"
description = "Blueprint services"
authors = ["NVIDIA"]
readme = "dfm/client/README.md"
packages = [{ include = "dfm/__init__.py" }, { include = "dfm/client" }, { include = "dfm/common" }, { include = "dfm/api" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
aiohttp = "^3.10.5"
requests = ">=2.31.0,<2.32.0"
pydantic = "^2.9.0"
dask = {extras = ["distributed"], version = "^2024.10.0"}

[tool.poetry.group.core]
optional = true

[tool.poetry.group.core.dependencies]
python = ">=3.12,<3.13"
aiohttp = "^3.10.5"
requests = ">=2.31.0,<2.32.0"
xarray = "^2024.7.0"
redis = "^5.0.8"
json-log-formatter = "^1.0"
websockets = "^13.0.1"
pillow = "^11.0.0"
pygrib = "^2.1.6"
pooch = "^1.8.2"
cfgrib = "^0.9.14.0"
scipy = "^1.14.1"
zarr = "^2.18.3"
cartopy = "^0.23.0"
watchfiles = "^0.24.0"
httpx = "^0.27.2"
opentelemetry-distro = "^0.48b0"
opentelemetry-semantic-conventions = "^0.48b0"
uvicorn = "^0.30.6"
boto3 = "^1.35.23"
fsspec = "^2024.9.0"
s3fs = "^2024.9.0"
h5netcdf = "^1.3.0"
cftime = "^1.6.4"
gcsfs = "^2024.9.0.post1"
netcdf4 = "^1.7.1.post2"
pydap = "^3.5"
geopandas = "^1.0.1"
affine = "^2.4.0"
rasterio = "^1.4.1"
shapely = "^2.0.6"
pyproj = "^3.7.0"
tilematrix = "^2023.12.0"
herbie-data = "^2024.8.0"
rioxarray = "^0.18.2"
metpy = "^1.6.3"
arcgis = "^2.4.0"
arcgis-mapping = "^4.30.0"
fastapi = "^0.115.11"
starlette = "^0.46.0"


[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
packaging = "^24.1"
fakeredis = { extras = ["json"], version = "^2.24.1" }
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
pytest-asyncio = "^0.24.0"
pylint = "^3.2.7"
bokeh = "^3.5.2"
mkdocs = "^1.6.1"
boto3 = "^1.35.23"
fsspec = "^2024.9.0"
s3fs = "^2024.9.0"
h5netcdf = "^1.3.0"
cftime = "^1.6.4"

scikit-image = "^0.24.0"
click = "^8.1.7"
sh = "^2.1.0"
mkdocstrings = {extras = ["python"], version = "^0.26.2"}
mkdocstrings-python = "^1.12.2"
jupyterlab = "^4.3.5"
pydantic = "^2.10.6"
ruamel-yaml = "^0.18.10"
gitpython = "^3.1.44"
pytest-repeat = "^0.9.3"
coloredlogs = "^15.0.1"
hvac = "^2.3.0"

[tool.dist]
version = "ea1"
output = "artifacts/dist"
sources = [
    'dfm',
    'k8s',
    'poetry.lock',
    'pyproject.toml',
    'examples/client/earth-2',
    'examples/minikube',
    'examples/build_images.sh',
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
